requirements.txt
discord.py
python-dotenv
together
pytz
